+++ src/Ectotherm.jl
@@ -1,0 +1,12 @@
+module Ectotherm
+
+using Unitful
+
+using Unitful: °, rad, °C, K, Pa, kPa, MPa, J, kJ, W, L, g, kg, cm, m, s, hr, d, mol, mmol, μmol, σ, R
+
+
+include("biophysics.jl") end
+include("GEOM.jl") end
+include("ectotherm.jl") end
+
+end
